+++ src/main/java/com/github/standobyte/jojo/client/render/armor/StoneMaskArmorClientExtensions.java
@@ -1,13 +1,34 @@
 package com.github.standobyte.jojo.client.render.armor;
 
+import com.github.standobyte.jojo.client.ModEntityTypeRenderers;
+import com.github.standobyte.jojo.client.render.armor.model.StoneMaskArmorModel;
+
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.model.HumanoidModel;
+import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.item.ArmorMaterial;
 import net.minecraft.world.item.ItemStack;
 import net.neoforged.neoforge.client.extensions.common.IClientItemExtensions;
 
 public class StoneMaskArmorClientExtensions implements IClientItemExtensions {
+	private StoneMaskArmorModel<LivingEntity> model;
+
 	@Override
-	public int getArmorLayerTintColor(ItemStack stack, LivingEntity entity, ArmorMaterial.Layer layer, int layerIdx, int fallbackColor) {
-		return 0;
+	public HumanoidModel<?> getHumanoidArmorModel(
+			LivingEntity livingEntity,
+			ItemStack itemStack,
+			EquipmentSlot equipmentSlot,
+			HumanoidModel<?> original) {
+		if (equipmentSlot != EquipmentSlot.HEAD) {
+			return original;
+		}
+		if (model == null) {
+			model = new StoneMaskArmorModel<>(
+					Minecraft.getInstance().getEntityModels()
+							.bakeLayer(
+									ModEntityTypeRenderers
+											.STONE_MASK_ARMOR));
+		}
+		return model;
 	}
 }