+++ src/main/java/com/github/standobyte/jojo/client/ui/screen_jojomenu/TabCategory.java
@@ -6,7 +6,6 @@ import java.util.function.Supplier;
 
 import javax.annotation.Nullable;
 
-import com.github.standobyte.jojo.client.ClientPowerCache;
 import com.github.standobyte.jojo.client.ui.utils.GuiIcon;
 import com.github.standobyte.jojo.powersystem.Power;
 import com.github.standobyte.jojo.powersystem.PowerClass;
@@ -39,7 +38,8 @@ public class TabCategory implements IJojoMenuTab {
 	
 	public boolean isActive() {
 		if (powerClass != null) {
-			Power<?> power = ClientPowerCache.getPower(powerClass);
+			Power<?> power =
+					JojoMenuTabs.getPowerForMenu(powerClass);
 			return power != null && power.hasPower() && (powerType == null || power.getPowerType() == powerType.get());
 		}
 		for (Tab tab : tabs) {