+++ src/main/java/com/github/standobyte/jojo/item/StoneMaskItem.java
@@ -1,6 +1,7 @@
 package com.github.standobyte.jojo.item;
 
 import com.github.standobyte.jojo.block.StoneMaskBlock;
+import com.github.standobyte.jojo.init.ModArmorMaterials;
 
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.component.DataComponents;
@@ -11,7 +12,7 @@ import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ArmorItem;
-import net.minecraft.world.item.ArmorMaterials;
+import net.minecraft.world.item.ArmorMaterial;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.BlockPlaceContext;
@@ -35,7 +36,7 @@ public class StoneMaskItem extends ArmorItem {
 	}
 
 	public StoneMaskItem(Properties properties, StoneMaskBlock block) {
-		super(ArmorMaterials.LEATHER, ArmorItem.Type.HELMET, properties);
+		super(ModArmorMaterials.STONE_MASK, ArmorItem.Type.HELMET, properties);
 		this.block = block;
 		BY_BLOCK.put(block, this);
 	}
@@ -71,6 +72,16 @@ public class StoneMaskItem extends ArmorItem {
 				itemId.getNamespace(),
 				"textures/armor/" + itemId.getPath()
 						+ suffix + ".png");
+	}
+
+	@Override
+	public ResourceLocation getArmorTexture(
+			ItemStack stack,
+			Entity entity,
+			EquipmentSlot slot,
+			ArmorMaterial.Layer layer,
+			boolean innerModel) {
+		return getArmorTexture(stack);
 	}
 
 	@Override